Added PendingOrder functions
Added PendingOrder functions and fixed a potential bug in the MarketTrade class where the total number of orders could potentially change while the code is looping through the order pool. Placed a call to the MQL4 OrdersTotal function in the for loop termination condition of several MarketTrade functions to correct this potential bug.